css/css-fonts/parsing/font-variant-numeric-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-variant-numeric with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-fonts/#font-variant-numeric-prop">
<meta name="assert" content="font-variant-numeric supports the full grammar 'normal | [ <numeric-figure-values> || <numeric-spacing-values> || <numeric-fraction-values> || ordinal || slashed-zero ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("font-variant-numeric", "normal");

test_valid_value("font-variant-numeric", "lining-nums");
test_valid_value("font-variant-numeric", "oldstyle-nums");

test_valid_value("font-variant-numeric", "proportional-nums");
test_valid_value("font-variant-numeric", "tabular-nums");

test_valid_value("font-variant-numeric", "diagonal-fractions");
test_valid_value("font-variant-numeric", "stacked-fractions");

test_valid_value("font-variant-numeric", "ordinal");

test_valid_value("font-variant-numeric", "slashed-zero");

test_valid_value("font-variant-numeric", "oldstyle-nums tabular-nums stacked-fractions ordinal slashed-zero");
// Blink seralizes with non-canonical order.
test_valid_value("font-variant-numeric", "slashed-zero ordinal stacked-fractions tabular-nums oldstyle-nums", "oldstyle-nums tabular-nums stacked-fractions ordinal slashed-zero");
</script>
</body>
</html>
